% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAQI-package.R
\docType{package}
\name{claiR-package}
\alias{claiR-package}
\title{Assessment of Air Cleanliness using Open Data}
\description{
Assessment of Air Cleanliness using Open Data
}
\details{
Access real-time data through the World Air Quality Index (WAQI) project's
API.
}
\author{
Florian Detsch
\cr
\cr
\emph{Maintainer:} Florian Detsch \email{fdetsch@web.de}
}
\keyword{package}
