% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory}
\alias{inventory}
\title{Query WAQI Stations Within Spatial Extent}
\usage{
inventory(x, token, ...)
}
\arguments{
\item{x}{A \code{sf} or \code{Spatial} object from which to extract spatial extent.}

\item{token}{\href{http://aqicn.org/faq/2015-09-18/map-web-service-real-time-air-quality-tile-api/}{API token ID}
for Air Quality Open Data Platform as \code{character}.}

\item{...}{Currently not used.}
}
\value{
Monitoring stations within spatial extent as \code{sf} object.
}
\description{
Access the WAQI API to query all available monitoring stations within a given
spatial extent.
}
\note{
The \strong{rgeos} package is required for finding out which hole belongs to
which exterior ring in \code{\link[sf]{st_as_sfc}} for \code{SpatialPolygons}.
}
\examples{
\dontrun{
deu = raster::getData(country = "DEU", level = 0)
stn = inventory(deu, token = keyring::key_get("waqi_api"))

library(mapview)
mapview(deu, alpha.regions = .25) + stn
}

}
