% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metainfo.R
\name{metainfo}
\alias{metainfo}
\title{Query WAQI Stations Metainformation}
\usage{
metainfo(x, token, uid = "uid", stock = NULL, sort = TRUE, ...)
}
\arguments{
\item{x}{WAQI stations as \code{sf} or \code{Spatial} object, usually derived
from \code{\link{inventory}}.}

\item{token}{\href{http://aqicn.org/faq/2015-09-18/map-web-service-real-time-air-quality-tile-api/}{API token ID}
for Air Quality Open Data Platform as \code{character}.}

\item{uid}{Column in 'x' with unique monitoring station ID as
\code{character}; defaults to "uid" which is the \href{http://aqicn.org/json-api/doc/}{API standard}.}

\item{stock}{A \code{data.frame} of already available metainformation created
during a previous function run, see 'Details'.}

\item{sort}{A \code{logical} determining whether the rows are to be re-arranged
by unique IDs.}

\item{...}{Additional arguments passed to \code{\link{makePSOCKcluster}}.}
}
\value{
Retrieved metainformation as \code{data.frame}.
}
\description{
Access the WAQI API to query metainformation for a given set of monitoring
stations.
}
\details{
Providing already available metainformation is a convenient way to process
either newly added monitoring stations or stations that have temporarily been
out of service during a previous function run only, and hence reduce
computation time significantly.
}
\examples{
\dontrun{
tkn = keyring::key_get("waqi_api")

deu = raster::getData(country = "DEU", level = 0)
stn = inventory(deu, token = tkn)

nfo = metainfo(stn, token = tkn)
head(nfo, 10)
}

}
\seealso{
\code{\link{inventory}}
}
