% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAQIControls.R
\name{preparation}
\alias{preparation}
\title{Preprocess WAQI Data}
\usage{
preparation(x)
}
\arguments{
\item{x}{WAQI data as \code{data.frame}.}
}
\value{
An adjusted \code{data.frame}.
}
\description{
Preprocess WAQI data, ie. declare missing values \code{NA}, reformat hourly
time offsets (sometimes ':' is not included, which is required by connected
functions, e.g. from \strong{lubridate}), and remove duplicate rows.
}
