% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readings.R
\name{readings}
\alias{readings}
\title{Query WAQI Stations Readings}
\usage{
readings(x, token, uid = "uid", dsn = NULL, ...)
}
\arguments{
\item{x, token, uid}{See \code{\link{metainfo}}.}

\item{dsn}{Optional target folder for automated \code{.fst} file output,
see also \code{\link[fst]{write_fst}}.}

\item{...}{Additional arguments passed to \code{\link{makePSOCKcluster}}.}
}
\value{
Retrieved readings as \code{data.frame}. Note that stations that could not be
found or were temporarily out of service are not included.
}
\description{
Access the WAQI API to query readings for a given set of monitoring stations.
}
\examples{
\dontrun{
## spatial extent: bavaria
deu = raster::getData(country = "DEU", level = 1, path = raster::tmpDir())
bay = subset(deu, NAME_1 == "Bayern")

## api token
tkn = keyring::key_get("waqi_api")

## query monitoring stations within spatial extent
stn = inventory(bay, token = tkn)

## query station readings
vls = readings(stn, token = tkn, names = parallel::detectCores() - 1)
vls
}

}
\seealso{
\code{\link{inventory}}
}
