% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieval.R
\name{retrieval}
\alias{retrieval}
\title{Query WAQI Stations Metainformation and Readings}
\usage{
retrieval(x, token, uid = "uid", stock = NULL, sort = TRUE, dsn = NULL, ...)
}
\arguments{
\item{x, token, uid, stock, sort, dsn, ...}{See \code{\link{metainfo}} and
\code{\link{readings}}.}
}
\value{
A \code{list} with retrieved (i) metainformation and (ii) readings, each as a
separate \code{data.frame}.
}
\description{
Access the WAQI API to query metainformation and readings for a given set of
monitoring stations.
}
\examples{
## see ?metainfo and ?readings

}
