% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAQI-package.R
\docType{package}
\name{WAQI-package}
\alias{WAQI-package}
\title{World Air Quality Index Project Data Retrieval}
\description{
World Air Quality Index Project Data Retrieval
}
\details{
Access real-time data through the World Air Quality Index (WAQI) project's
API.
}
\author{
Florian Detsch
\cr
\cr
\emph{Maintainer:} Florian Detsch \email{fdetsch@web.de}
}
\keyword{package}
