% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAQI-package.R
\docType{data}
\name{WAQIstations}
\alias{WAQIstations}
\title{WAQI Stations in Germany}
\format{
\code{sf data.frame} with 210 features and 2 fields covering a
spatial extent from 6.094444, 47.515447, 14.638056, 53.638190 (xmin, ymin,
xmax, ymax; EPSG:4326).
}
\source{
\url{http://aqicn.org/api/de/}
}
\description{
World Air Qualilty Index (WAQI) project monitoring stations
located in Germany.
}
\details{
Data available through the Air Quality Programmatic API (see Source).
}
